+++ config/settings/base.py
@@ -7,11 +7,16 @@ https://docs.djangoproject.com/en/dev/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/dev/ref/settings/
 """
+
 import environ
+import sys
 
 SITE_PACKAGE = "varfish"
 ROOT_DIR = environ.Path(__file__) - 3  # (varfish/config/settings/base.py - 3 = varfish/)
 APPS_DIR = ROOT_DIR.path("varfish")
+
+# Check whether we are running tsts (this is important to use models and not materialized views in tests).
+IS_TESTING = len(sys.argv) > 1 and sys.argv[1] == 'test'
 
 # Load operating system environment variables and then prepare to use them
 env = environ.Env()