+++ variants/models_support.py
@@ -18,7 +18,7 @@ from variants.models import (
     SmallVariant,
     SmallVariantComment,
     SmallVariantFlags,
-    SmallVariantQuery,
+    SmallVariantSummary,
 )
 from frequencies.models import GnomadExomes, GnomadGenomes, Exac, ThousandGenomes
 from variants.forms import (
@@ -713,6 +713,10 @@ class FilterQueryRenderFieldsMixin(JoinDbsnpAndHgncMixin):
                 SmallVariant.sa.gnomad_exomes_homozygous,
                 SmallVariant.sa.gnomad_genomes_homozygous,
                 SmallVariant.sa.thousand_genomes_homozygous,
+                SmallVariant.sa.exac_heterozygous,
+                SmallVariant.sa.gnomad_exomes_heterozygous,
+                SmallVariant.sa.gnomad_genomes_heterozygous,
+                SmallVariant.sa.thousand_genomes_heterozygous,
                 SmallVariant.sa.genotype,
                 SmallVariant.sa.case_id,
                 SmallVariant.sa.in_clinvar,
@@ -1011,8 +1015,88 @@ class FilterQueryFlagsCommentsMixin:
         return or_(*terms)
 
 
+class FilterInHouseCountsMixin:
+    """Join in-house variants to the results and threshold on the number of heterozygous/homozygous variants."""
+
+    def _build_stmt(self, kwargs):
+        """Override statement building to add the join with Clinvar information."""
+        inner = super()._build_stmt(kwargs)
+        return self._extend_stmt_inhouse_db(inner, kwargs)
+
+    def _extend_stmt_inhouse_db(self, inner, kwargs):
+        """Extend the inner statement and augment inhouse count information."""
+        inner = inner.alias("inhouse_inner")
+        middle = (
+            select(
+                [
+                    *inner.c,
+                    func.coalesce(func.sum(SmallVariantSummary.sa.count_hom_ref), 0).label(
+                        "inhouse_hom_ref"
+                    ),
+                    func.coalesce(func.sum(SmallVariantSummary.sa.count_het), 0).label(
+                        "inhouse_het"
+                    ),
+                    func.coalesce(func.sum(SmallVariantSummary.sa.count_hom_alt), 0).label(
+                        "inhouse_hom_alt"
+                    ),
+                    func.coalesce(func.sum(SmallVariantSummary.sa.count_hemi_ref), 0).label(
+                        "inhouse_hemi_ref"
+                    ),
+                    func.coalesce(func.sum(SmallVariantSummary.sa.count_hemi_alt), 0).label(
+                        "inhouse_hemi_alt"
+                    ),
+                    func.coalesce(
+                        func.sum(
+                            SmallVariantSummary.sa.count_het
+                            + SmallVariantSummary.sa.count_hom_alt
+                            + SmallVariantSummary.sa.count_hemi_alt
+                        ),
+                        0,
+                    ).label("inhouse_carriers"),
+                ]
+            )
+            .select_from(
+                inner.outerjoin(
+                    SmallVariantSummary.sa.table,
+                    and_(
+                        SmallVariantSummary.sa.release == inner.c.release,
+                        SmallVariantSummary.sa.chromosome == inner.c.chromosome,
+                        SmallVariantSummary.sa.position == inner.c.position,
+                        SmallVariantSummary.sa.reference == inner.c.reference,
+                        SmallVariantSummary.sa.alternative == inner.c.alternative,
+                    ),
+                )
+            )
+            .group_by(*inner.c)
+            .alias("inhouse_middle")
+        )
+        stmt = select(middle.c).select_from(middle).where(self._where_inhouse_db(kwargs, middle))
+        return self._add_trailing(stmt, kwargs)
+
+    def _where_inhouse_db(self, kwargs, stmt):
+        """Build WHERE clause for the query based on select het/hom counts in inhouse DB."""
+        terms = []
+        if kwargs.get("inhouse_enabled"):
+            if kwargs.get("inhouse_heterozygous") is not None:
+                terms.append(stmt.c.inhouse_het <= kwargs.get("inhouse_heterozygous"))
+            if kwargs.get("inhouse_homozygous") is not None:
+                terms.append(
+                    (stmt.c.inhouse_hom_alt + stmt.c.inhouse_hemi_alt)
+                    <= kwargs.get("inhouse_homozygous")
+                )
+            if kwargs.get("inhouse_carriers") is not None:
+                terms.append(
+                    (stmt.c.inhouse_het + stmt.c.inhouse_hom_alt + stmt.c.inhouse_hemi_alt)
+                    <= kwargs.get("inhouse_carriers")
+                )
+        return and_(*terms)
+
+
 class JoinAndQueryCommonAdditionalTables(
-    FilterQueryFlagsCommentsMixin, FilterQueryHgmdMixin, FilterQueryClinvarDetailsMixin
+    FilterQueryFlagsCommentsMixin,
+    FilterQueryHgmdMixin,
+    FilterQueryClinvarDetailsMixin,
+    FilterInHouseCountsMixin,
 ):
     """Join to common common additional tables and query.
 
@@ -1069,7 +1153,10 @@ class ExportTableFileFilterQuery(
 
 
 class ExportVcfFileFilterQuery(
-    OrderByChromosomalPositionMixin, BaseTableQueriesMixin, SingleCaseFilterQueryBase
+    OrderByChromosomalPositionMixin,
+    FilterInHouseCountsMixin,
+    BaseTableQueriesMixin,
+    SingleCaseFilterQueryBase,
 ):
     """Run filter query for TSV file with minimal information only for performance.